Add Storage, Parser, TaskList OOP
